+++ srcs/parser/quoting/pre_parsing_work.c
@@ -1,3 +1,15 @@
+/* ************************************************************************** */
+/*                                                                            */
+/*                                                        :::      ::::::::   */
+/*   pre_parsing_work.c                                 :+:      :+:    :+:   */
+/*                                                    +:+ +:+         +:+     */
+/*   By: rbednar <rbednar@student.21-school.ru>     +#+  +:+       +#+        */
+/*                                                +#+#+#+#+#+   +#+           */
+/*   Created: 2020/07/25 15:49:13 by rbednar           #+#    #+#             */
+/*   Updated: 2020/08/01 15:23:10 by rbednar          ###   ########.fr       */
+/*                                                                            */
+/* ************************************************************************** */
+
 #include "shell42.h"
 #include "parser.h"
 
@@ -6,7 +18,7 @@
 ** g_techline.line and resizing g_cmd_size and g_techline.len.
 */
 
-int		ft_reglue(size_t *i, int num, t_ltree *sub)
+int		ft_reglue(int *i, int num, t_ltree *sub)
 {
 	size_t	z;
 	size_t	size;
@@ -27,7 +39,7 @@ int		ft_reglue(size_t *i, int num, t_ltree *sub)
 	return (0);
 }
 
-int		pre_parsing_back(size_t *i, t_ltree *sub)
+int		pre_parsing_back(int *i, t_ltree *sub)
 {
 	char	*end;
 
@@ -44,7 +56,7 @@ int		pre_parsing_back(size_t *i, t_ltree *sub)
 	return (0);
 }
 
-int		pre_parsing_andor_pipe(size_t *i, t_ltree *sub)
+int		pre_parsing_andor_pipe(int *i, t_ltree *sub)
 {
 	char	*end;
 
@@ -56,7 +68,7 @@ int		pre_parsing_andor_pipe(size_t *i, t_ltree *sub)
 		if (end[*i] == PIPE && end[*i + 1] != PIPE)
 			*i += 1;
 		else
-			*i += 2;		
+			*i += 2;
 		while (end[*i] == SPACE)
 			(*i)++;
 		if (end[*i] == ENTER)
@@ -68,7 +80,7 @@ int		pre_parsing_andor_pipe(size_t *i, t_ltree *sub)
 	return (0);
 }
 
-int		pre_parsing_squote(size_t *i, t_ltree *sub)
+int		pre_parsing_squote(int *i, t_ltree *sub)
 {
 	char	*end;
 
@@ -99,11 +111,11 @@ int		pre_parsing_squote(size_t *i, t_ltree *sub)
 int		pre_parsing_cut_glue(t_ltree *sub)
 {
 	char	*end;
-	size_t	i;
+	int		i;
 
 	i = 0;
 	end = sub->l_tline.line;
-	while (i < sub->l_tline.len)
+	while ((size_t)i < sub->l_tline.len)
 	{
 		if (end[i] == DQUOTE)
 		{
@@ -122,6 +134,5 @@ int		pre_parsing_cut_glue(t_ltree *sub)
 			ft_reglue(&i, 1, sub);
 		i++;
 	}
-		// print_techline(sub->l_cmd, sub->l_tline.line, sub->l_tline.len);
 	return (0);
 }
